+++ vcfstats/__main__.py
@@ -1,0 +1,5 @@
+"""Main entrance for `python -m vcfstats`"""
+
+from .cli import main
+
+main()